psutil-mqtt/requirements.txt
nvidia-ml-py3==7.352.0
paho-mqtt==1.6.1
psutil==5.9.2
jq==1.3.0